Consume chainDataGet route in views
