--- packages/admin-ui/src/services/chainData/actions.ts
@@ -1,6 +1,0 @@
-import { createAction } from "@reduxjs/toolkit";
-import { ChainData } from "types";
-
-// Service > chainData
-
-export const updateChainData = createAction<ChainData[]>("chainData/update");
